servers/semrush-appcenter/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcparmory-semrush-appcenter"
version = "1.0.2"
description = "Manage subscriptions, users, and send event notifications for Semrush apps"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=3.2.4",
    "httpx>=0.27.0,<1.0.0",
    "mcp",
    "pydantic>=2.0.0,<3.0.0",
    "PyJWT[crypto]>=2.8.0,<3.0.0",
    "python-dotenv>=1.0.0,<2.0.0"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]

[project.scripts]
mcparmory-semrush-appcenter = "server:main"

[project.urls]
Homepage = "https://mcparmory.com"
Repository = "https://github.com/mcparmory/registry"

[tool.hatch.build.targets.wheel]
packages = ["."]
exclude = [
    "*.json",
    ".env*",
    ".mcp.json",
    "Dockerfile",
    "requirements.txt",
    "dist",
]
